Feat : postingLike 예외처리 추가

+++ server/src/main/java/com/main19/server/posting/like/service/PostingLikeService.java
@@ -64,6 +64,11 @@ public class PostingLikeService {
 		}
 
 		PostingLike findPostingLike = findVerifiedPostingLike(postingLikeId);
+
+		if(findPostingLike.getPostingLikeCount() == 0) {
+			throw new BusinessLogicException(ExceptionCode.POSTING_LIKE_NOT_FOUND);
+		}
+
 		findPostingLike.getPosting().deleteLikeCount();
 
 		postingLikeRepository.delete(findPostingLike);
